+++ imio/dms/mail/browser/actionspanel.py
@@ -3,7 +3,6 @@ from imio.actionspanel.browser.viewlets import ActionsPanelViewlet
 from imio.actionspanel.browser.views import ActionsPanelView
 from imio.dms.mail.dmsmail import filter_dmsincomingmail_assigned_users
 from imio.helpers.cache import get_plone_groups_for_user
-from imio.helpers.security import check_zope_admin
 from plone import api
 from plone.memoize import ram
 from Products.CMFPlone.interfaces import IHideFromBreadcrumbs
@@ -187,8 +186,7 @@ class SigningFieldsetActionsPanelView(ActionsPanelView):
 
     def __init__(self, context, request):
         super(SigningFieldsetActionsPanelView, self).__init__(context, request)
-        self.ACCEPTABLE_ACTIONS = ["edit", "approvals"]
-        self.SECTIONS_TO_RENDER += ("renderSigningAnnotationInfo",)
+        self.ACCEPTABLE_ACTIONS = ["edit", "approvals", "session-annotation-info"]
 
     @property
     def fieldset(self):
@@ -197,11 +195,6 @@ class SigningFieldsetActionsPanelView(ActionsPanelView):
     def renderEdit(self):
         if self.showEdit and self.mayEdit():
             return ViewPageTemplateFile("templates/fieldset_actions_panel_edit.pt")(self)
-        return ""
-
-    def renderSigningAnnotationInfo(self):
-        if check_zope_admin():
-            return ViewPageTemplateFile("templates/fieldset_actions_panel_signing_annotation_info.pt")(self)
         return ""
 
     @ram.cache(actionspanelview_cachekey)